fix/view 파일 분리/articles카테고리 추가/ articles:디테일 프론트/ articles:index 프론트
