+++ articles/models.py
@@ -4,6 +4,16 @@ from django.db import models
 from django.conf import settings 
 from datetime import datetime, timedelta,timezone
 # Create your models here.
+class CategorySelect(models.IntegerChoices):
+    자유 = 1, '자유'
+    유머 = 2, '유머'
+    팬아트 = 3, '팬아트'
+    유저찾기 = 4, '유저찾기'
+    유저뉴스 = 5, '유저뉴스'
+    팁과노하우 = 6, '팁과노하우'
+    기획 = 7, '기획'
+    사건사고 = 8, '사건사고'
+
 class Article(models.Model):
     title = models.CharField(max_length=20)
     content = models.TextField()
@@ -14,6 +24,7 @@ class Article(models.Model):
         processors=[ResizeToFill(100, 80)],
         format='JPEG',
         options={'quality': 80},)
+    category = models.IntegerField(default=CategorySelect.자유,choices=CategorySelect.choices)
     like_user = models.ManyToManyField(
         settings.AUTH_USER_MODEL, related_name="like_post"
     )
