+++ articles/templates/articles/index.html
@@ -7,16 +7,20 @@
 <style>
     a{
         text-decoration:none;
+        color:black;
+    }
+    a:active{
+        background-color:grey;
+        color:green;
     }
 </style>
 <!-- 게시판 -->
 
 <div class="row my-3">
     <!-- 왼쪽 어사이드 -->
-    <div class="d-none d-lg-block col-lg-3 border border-3 my-3 rounded-3" style="max-height:300px; position:sticky; top:0;">
+    {% if request.user.is_authenticated %}
+    <div class="d-none d-lg-block col-lg-3 border border-3 my-3 rounded-3" style="max-height:600px; position:sticky; top:0;">
         <div class="my-2 container">
-        {% if request.user.is_authenticated %}
-        
             <p>
                 <img style="width: 20%; height:20%;" src="{% static 'images/logo.png' %}" alt=""> <span>{{ user.username }}</span>
                 <span>
@@ -25,15 +29,50 @@
                     </div>
                 </span>
             </p>
-            <a href="?sort=3" class="btn btn-outline-success">내가 쓴 글</a>
-            <a href="?sort=4" class="btn btn-outline-success">내 와드</a>
-            <a href="?sort=5" class="btn btn-outline-success">내 좋아요</a>
+            <hr>
+            <div class="d-flex justify-content-between my-2">
+                <a href="?sort=3" class="btn btn-outline-success">내가 쓴 글</a>
+                <a href="?sort=4" class="btn btn-outline-success">내 와드</a>
+                <a href="?sort=5" class="btn btn-outline-success">좋아요</a>
+            </div>
+            <hr>
             <a href="{% url 'articles:create' %}" class="btn btn-success w-100">글쓰기</a>
-            {% else %}
-                <a href="{% url 'accounts:login' %}" class="btn btn-primary w-100">로그인</a>
-            {% endif %}
+            <hr>
+            <div class="d-flex flex-column">
+                <a name="tag1" class="my-1 small" href="?sort=6">자유</a>
+                <a class="my-1 small"href="?sort=7">유머</a>
+                <a class="my-1 small"href="?sort=10">팬아트</a>
+                <hr>
+                <a class="my-1 small"href="?sort=8">유저찾기</a>
+                <a class="my-1 small"href="?sort=12">유저뉴스</a>
+                <a class="my-1 small"href="?sort=13">팁과노하우</a>
+                <hr>
+                <a class="my-1 small"href="?sort=11">기획</a>
+                <a class="my-1 small"href="?sort=9">사건사고</a>
+            </div>
         </div>
     </div>
+    {% else %}
+    <div class="d-none d-lg-block col-lg-3 border border-3 my-3 rounded-3" style="max-height:400px; position:sticky; top:0;">
+        <div class="my-2 container">
+            <a href="{% url 'accounts:login' %}" class="btn btn-primary w-100 mt-2">로그인</a>
+            <hr>
+            <div class="d-flex flex-column">
+                <a class="my-1 small"href="?sort=6">자유</a>
+                <a class="my-1 small"href="?sort=7">유머</a>
+                <a class="my-1 small"href="?sort=10">팬아트</a>
+                <hr>
+                <a class="my-1 small"href="?sort=8">유저찾기</a>
+                <a class="my-1 small"href="?sort=12">유저뉴스</a>
+                <a class="my-1 small"href="?sort=13">팁과노하우</a>
+                <hr>
+                <a class="my-1 small"href="?sort=11">기획</a>
+                <a class="my-1 small"href="?sort=9">사건사고</a>
+            </div>
+        </div>
+    </div>
+            {% endif %}
+  
     <!-- 게시판 내용 -->
     <div class="col-lg-9 col-md-12">
         <div class="border border-3 d-flex flex-column my-3 rounded-3">
@@ -149,3 +188,9 @@
     </div>
 </div>
 {% endblock content %}
+<script src="https://code.jquery.com/jquery-3.6.1.min.js" integrity="sha256-o88AwQnZB+VDvE9tvIXrMQaPlFFSUTR+nldQm1LuPXQ=" crossorigin="anonymous"></script>
+<script>
+    $("a").click(function(){
+        toggleClass(".active-color");
+    })
+</script>
